Logging under the same header in stories
