+++ packages/react-components/lib/date-time-picker.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 import {
   DateTimePickerProps,
   KeyboardDateTimePicker,
@@ -7,22 +6,7 @@ import {
 } from '@material-ui/pickers';
 import MomentUtils from '@date-io/moment';
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    container: {
-      display: 'flex',
-      flexWrap: 'wrap',
-    },
-    textField: {
-      marginLeft: theme.spacing(1),
-      marginRight: theme.spacing(1),
-      width: 200,
-    },
-  }),
-);
-
 export default function DateAndTimePickers(props: DateTimePickerProps): React.ReactElement {
-  const classes = useStyles();
   const { name, label, value, error, ...rest } = props;
   return (
     <MuiPickersUtilsProvider utils={MomentUtils}>
